Some overflow fixes

+++ src/learner.c
@@ -661,9 +661,9 @@ void lea_local_fiber(struct fbr_context *fiber_context, void *_arg)
 	struct fbr_mutex mutex;
 	struct fbr_cond_var *cond;
 	struct acc_archive_record *arecords;
-	unsigned x;
-	unsigned count;
-	const unsigned read_batch = 1000;
+	uint64_t x;
+	uint64_t count;
+	const uint64_t read_batch = 1000;
 	struct lea_local_state_i local_state;
 	struct fbr_destructor dtor = FBR_DESTRUCTOR_INITIALIZER;
